+++ Aman-Srivastava/src/Programs/Aman-Srivastava-Project/Project-Practice/main.py
@@ -1,4 +1,4 @@
-from src.utils.data_extraction import Plot_Class
+from src.data_extraction.knolder_employee_score_extraction.data_extraction import Plot_Class
 
 if __name__ == "__main__":
     read_input = int(input("Enter Your Knolder ID: "))
