+++ Aman-Srivastava/src/Programs/Aman-Srivastava-Project/Project-Practice/requirements.txt
@@ -1,4 +1,4 @@
 pandas==1.5.3
 matplotlib==3.5.1
 numpy==1.24.2
-
+pytest~=7.3.1
