+++ Aman-Srivastava/src/Programs/Aman-Srivastava-Project/Project-Practice/src/utils/constants.py
@@ -1,1 +1,7 @@
-dataset_path = "/home/knoldus/Aman/Project-Practice/DataSets/individual_contribution.csv"
+from pathlib import Path
+
+p = Path(__file__).resolve().parent.parent
+
+method_path = p / "data_extraction"
+files_path = method_path / "csv_data"
+dataset_path = files_path / "individual_contribution.csv"
